+++ src/main/java/dev/saintho/mytly/event/dto/UrlCreateEvent.java
@@ -1,6 +1,6 @@
 package dev.saintho.mytly.event.dto;
 
-import dev.saintho.mytly.entity.Url;
+import dev.saintho.mytly.domain.entity.Url;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
