+++ src/main/java/dev/saintho/mytly/repository/DailyEngagementRepository.java
@@ -2,7 +2,7 @@ package dev.saintho.mytly.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import dev.saintho.mytly.entity.DailyEngagement;
+import dev.saintho.mytly.domain.entity.DailyEngagement;
 
 public interface DailyEngagementRepository extends JpaRepository<DailyEngagement, Long> {
 }
