+++ src/main/java/dev/saintho/mytly/repository/RefererEngagementRepository.java
@@ -2,7 +2,7 @@ package dev.saintho.mytly.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import dev.saintho.mytly.entity.RefererEngagement;
+import dev.saintho.mytly.domain.entity.RefererEngagement;
 
 public interface RefererEngagementRepository extends JpaRepository<RefererEngagement, Long> {
 }
