+++ src/main/java/dev/saintho/mytly/repository/UrlRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import dev.saintho.mytly.entity.Url;
+import dev.saintho.mytly.domain.entity.Url;
 
 public interface UrlRepository extends JpaRepository<Url, Long> {
 	Optional<Url> findByShortened(String shortenend);
