+++ src/main/java/dev/saintho/mytly/service/UrlService.java
@@ -9,10 +9,10 @@ import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import dev.saintho.mytly.dto.command.UrlDeleteByShortenedCommand;
-import dev.saintho.mytly.dto.command.UrlShortCommand;
-import dev.saintho.mytly.dto.query.UrlRedirectQuery;
-import dev.saintho.mytly.entity.Url;
+import dev.saintho.mytly.web.v1.dto.command.UrlDeleteByShortenedCommand;
+import dev.saintho.mytly.web.v1.dto.command.UrlShortCommand;
+import dev.saintho.mytly.web.v1.dto.query.UrlRedirectQuery;
+import dev.saintho.mytly.domain.entity.Url;
 import dev.saintho.mytly.event.dto.UrlCreateEvent;
 import dev.saintho.mytly.exception.MytlyException;
 import dev.saintho.mytly.generator.url.UrlGenerator;
